reorganizing fields namespace
