--- src/Frozennode/Administrator/Fields/Color.php
@@ -1,6 +1,0 @@
-<?php
-namespace Frozennode\Administrator\Fields;
-
-class Color extends Text {
-
-}
