--- src/Frozennode/Administrator/Fields/Relationships/HasOne.php
@@ -1,6 +1,0 @@
-<?php
-namespace Frozennode\Administrator\Fields\Relationships;
-
-class HasOne extends HasOneOrMany {
-
-}
