+++ src/Frozennode/Administrator/Fields/Types/Color.php
@@ -1,0 +1,5 @@
+<?php namespace Frozennode\Administrator\Fields\Types;
+
+class Color extends Text {
+
+}
