+++ src/Frozennode/Administrator/Fields/Types/Number.php
@@ -1,5 +1,4 @@
-<?php
-namespace Frozennode\Administrator\Fields;
+<?php namespace Frozennode\Administrator\Fields\Types;
 
 class Number extends Field {
