+++ src/Frozennode/Administrator/Fields/Types/Relationships/HasOne.php
@@ -1,0 +1,5 @@
+<?php namespace Frozennode\Administrator\Fields\Types\Relationships;
+
+class HasOne extends HasOneOrMany {
+
+}
